About to fix motor library

+++ main/utils.c
@@ -1,7 +1,7 @@
 #include "utils.h"
 
 int32_t mod(int32_t x, int32_t m){
-    int16_t r = x % m;
+    int32_t r = x % m;
     return r < 0 ? r + m : r;
 }
